Moved board stuff to separate file
Prevents having to write duplicate code
